logs-microservice/requirements.txt
Flask==2.0.1
azure-storage-file-share==12.10.0
Werkzeug==2.2.2